LayoutTests/imported/w3c/web-platform-tests/css/css-values/inherit-function-invalidation-expected.txt


FAIL inherit() invalidates when values changes on parent element, inherited assert_equals: expected "2" but got "auto"
FAIL inherit() invalidates when values changes on parent element, inherited, deeper assert_equals: expected "2" but got "auto"
FAIL inherit() invalidates when values changes on parent element, non-inherited assert_equals: expected "2" but got "auto"
